Move datastore initialization to main.py for use in various locations
